requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
plotly==5.18.0
PyPDF2==3.0.1
pdfplumber==0.10.3
Pillow==10.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2
openpyxl==3.1.2
python-multipart==0.0.6
httpx==0.25.2
